Refactor: BindingError Response JSON type으로 변경
